Move frame handling to separate class
