+++ src/main/java/de/retest/web/RecheckSeleniumAdapter.java
@@ -8,7 +8,6 @@ import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Predicate;
@@ -73,7 +72,9 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 		final RootElement lastChecked =
 				convert( tagMapping, driver.getCurrentUrl(), driver.getTitle(), shoot( driver ) );
 
-		addChildrenFromFrames( driver, cssAttributes, lastChecked );
+		final FrameConverter frameConverter =
+				new FrameConverter( getQueryJS(), retestIdProvider, attributesProvider, defaultValueFinder );
+		frameConverter.addChildrenFromFrames( driver, cssAttributes, lastChecked );
 
 		if ( driver instanceof UnbreakableDriver ) {
 			((UnbreakableDriver) driver).setLastActualState( lastChecked );
@@ -89,43 +90,6 @@ public class RecheckSeleniumAdapter implements RecheckAdapter {
 		logger.info( "Checking website {} with {} elements.", url, mapping.size() );
 		return new PeerConverter( retestIdProvider, attributesProvider, mapping, title, screenshot, defaultValueFinder )
 				.convertToPeers();
-	}
-
-	private void addChildrenFromFrames( final WebDriver driver, final Set<String> cssAttributes,
-			final RootElement lastChecked ) {
-		final JavascriptExecutor jsExecutor = (JavascriptExecutor) driver;
-		final List<Element> frames =
-				de.retest.web.selenium.By.findElements( lastChecked.getContainedElements(), isFrame );
-
-		logger.debug( "Found {} frames, getting data per frame.", frames.size() );
-		for ( final Element frame : frames ) {
-			final String frameId = frame.getIdentifyingAttributes().get( "id" );
-			if ( frameId == null ) {
-				// TODO Implement handling e.g. via name, XPaht, etc.
-				logger.error( "Cannot retrieve frame with ID null from {}.", frame );
-				continue;
-			}
-			try {
-				logger.debug( "Switching to frame with ID {}.", frameId );
-				driver.switchTo().frame( frameId );
-				final String framePath = frame.getIdentifyingAttributes().getPath();
-				@SuppressWarnings( "unchecked" )
-				final PathsToWebDataMapping mapping = new PathsToWebDataMapping( framePath,
-						(Map<String, Map<String, Object>>) jsExecutor.executeScript( getQueryJS(), cssAttributes ) );
-				final RootElement frameContent = new PeerConverter( retestIdProvider, attributesProvider, mapping,
-						"frame-" + frameId, null, defaultValueFinder ) {
-					@Override
-					protected boolean isRoot( final String parentPath ) {
-						// handle trailing slashes...
-						return framePath.equals( parentPath.replaceAll( "/$", "" ) );
-					}
-				}.convertToPeers();
-				frame.addChildren( frameContent.getContainedElements() );
-			} catch ( final Exception e ) {
-				logger.error( "Exception retrieving data content of frame with ID {}.", frameId, e );
-			}
-			driver.switchTo().defaultContent();
-		}
 	}
 
 	private String getQueryJS() {
